Fix update condition

+++ src/Media/JPEGFile.cpp
@@ -535,7 +535,7 @@ void Media::JPEGFile::WriteJPGBuffer(IO::Stream *stm, const UInt8 *jpgBuff, UOSI
 					WriteInt16(&exifBuff[12], 42);
 					WriteInt32(&exifBuff[14], 8);
 					k = 8;
-					l = endOfst + 8;
+					l = (UInt32)(endOfst + 8);
 					exif->ToExifBuff(&exifBuff[10], &k, &l);
 					stm->Write(exifBuff, exifSize + 18);
 					MemFree(exifBuff);
